Add files via upload
The Seed list filters for close monuments, close outposts, and close villages (that are not spruce villages because they are horrible for raids, never raid them unless you have no other option in an actual run). The datapack itself is self-explanatory for the most part, the only thing you need to know is you are given a "Free Nether Portal" item, which places nether portals for you, as in an actual run's end game, you would likely already have a nether enter/blind into the stronghold. Speaking of that, I spawn a stronghold's portal room (portal already filled of course) in front of you if you blind from the nether into the overworld, at least 1280 blocks out (since these are in the ring).